+++ backend/__tests__/transportationForm.test.ts
@@ -1,6 +1,6 @@
 import request from "supertest";
 import app from "../index";
-import Application from "../models/ApplicationCurrentYear";
+import Application from "../models/Application";
 import { TRANSPORTATION_STATUS, TRANSPORTATION_TYPE } from '../constants';
 import lolex from "lolex";
