+++ backend/models/Application.ts
@@ -1,56 +1,31 @@
-import mongoose from "mongoose";
-import { Model, Schema } from "mongoose";
-import { IApplication } from "./Application.d"
-import applicationInfoSchema from "./applicationInfoSchema";
-import adminInfoSchema from "./adminInfoSchema";
-import reviewSchema from "./reviewSchema";
-import { STATUS, TRANSPORTATION_STATUS } from "../constants";
-import {values} from "lodash";
-import transportationInfoSchema from "./transportationInfoSchema";
-import s3FilePlugin from "../utils/file_plugin";
+import mongoose, { Model } from "mongoose";
+import { applicationSchema } from "./ApplicationAnyYear";
+import { IApplication } from "./Application.d";
+import { HACKATHON_YEAR_STRING } from "../constants";
 
-export const applicationSchema: Schema = new mongoose.Schema({
-    // user id is _id.
-    "_id": String,
-    "forms": { // can only be modified by user/editors
-        "application_info": applicationInfoSchema,
-        "transportation": transportationInfoSchema
-    },
-    "admin_info": adminInfoSchema, // Only editable by admin.
-    "reviews": [reviewSchema], // each review can only be modified by the reviewer who made it.
-    "user": {
-        "email": String
-    },
-    "status": {
-        type: String,
-        default: STATUS.INCOMPLETE,
-        enumValues: values(STATUS)
-    },
-    "transportation_status": {
-        type: String,
-        default: null,
-        enumValues: values(TRANSPORTATION_STATUS)
-    },
-    "type": {
-        type: String,
-        enumValues: ["is", "oos", "stanford"]
-    },
-    "location": {
-        type: String,
-        enumValues: ["Outside USA", "Alabama", "Alaska", "American Samoa", "Arizona", "Arkansas", "California", "Colorado", "Connecticut", "Delaware", "District Of Columbia", "Florida", "Georgia", "Guam", "Hawaii", "Idaho", "Illinois", "Indiana", "Iowa", "Kansas", "Kentucky", "Louisiana", "Maine", "Maryland", "Massachusetts", "Michigan", "Minnesota", "Mississippi", "Missouri", "Montana", "Nebraska", "Nevada", "New Hampshire", "New Jersey", "New Mexico", "New York", "North Carolina", "North Dakota", "Ohio", "Oklahoma", "Oregon", "Pennsylvania", "Rhode Island", "South Carolina", "South Dakota", "Tennessee", "Texas", "Utah", "Vermont", "Virginia", "Washington", "West Virginia", "Wisconsin", "Wyoming", "US Territories"]
-    },
-    "sponsor_optout": {
-        type: Boolean
-    },
-    "toDelete": {
-        type: Boolean
-    },
-    "year": {
-        type: String
-    }
+const applicationCurrentYearSchema = applicationSchema.clone();
+
+applicationCurrentYearSchema.pre('aggregate', function (this: mongoose.Aggregate<IApplication>) {
+    this.pipeline().unshift({
+        "$match": {
+            year: HACKATHON_YEAR_STRING,
+        }
+    })
 });
 
-applicationSchema.plugin(s3FilePlugin);
+applicationCurrentYearSchema.pre('find', projectCurrentYear);
+applicationCurrentYearSchema.pre('findOne', projectCurrentYear);
 
-const model: Model<IApplication> = mongoose.model("Application", applicationSchema, "Application");
+export function projectCurrentYear(this: mongoose.Query<IApplication>) {
+    let query = this.getQuery();
+    this.setQuery({ ...query, year: HACKATHON_YEAR_STRING });
+}
+
+async function setCurrentYear(this: mongoose.Document) {
+    if (!(this as any).year) {
+        (this as any).year = HACKATHON_YEAR_STRING;
+    }
+}
+
+const model: Model<IApplication> = mongoose.model("ApplicationCurrentYear", applicationSchema, "Application");
 export default model;
