+++ backend/routes/user_list.ts
@@ -1,4 +1,4 @@
-import Application from "../models/ApplicationCurrentYear";
+import Application from "../models/Application";
 import { Request, Response } from 'express';
 import { getGenericList } from "./common";
